+++ src/main/kotlin/com/interrupt/server/auth/dto/login/SignInResponse.kt
@@ -1,4 +1,4 @@
-package com.interrupt.server.member.dto.login
+package com.interrupt.server.auth.dto.login
 
 data class SignInResponse(
     val accessToken: String,
