+++ src/main/kotlin/com/interrupt/server/auth/dto/refresh/TokenRefreshRequest.kt
@@ -1,4 +1,4 @@
-package com.interrupt.server.auth.dto.request
+package com.interrupt.server.auth.dto.refresh
 
 data class TokenRefreshRequest(
     val refreshToken: String,
