+++ src/main/kotlin/com/interrupt/server/auth/service/AuthenticationService.kt
@@ -1,15 +1,15 @@
 package com.interrupt.server.auth.service
 
 import com.interrupt.server.auth.config.JwtProperties
-import com.interrupt.server.auth.dto.request.TokenRefreshRequest
-import com.interrupt.server.auth.dto.response.TokenRefreshResponse
+import com.interrupt.server.auth.dto.refresh.TokenRefreshRequest
+import com.interrupt.server.auth.dto.refresh.TokenRefreshResponse
 import com.interrupt.server.auth.entity.AuthenticationCredentials
 import com.interrupt.server.auth.entity.TokenCache
 import com.interrupt.server.auth.repository.TokenRedisRepository
 import com.interrupt.server.common.exception.ErrorCode
 import com.interrupt.server.common.exception.InterruptServerException
-import com.interrupt.server.member.dto.login.SignInRequest
-import com.interrupt.server.member.dto.login.SignInResponse
+import com.interrupt.server.auth.dto.login.SignInRequest
+import com.interrupt.server.auth.dto.login.SignInResponse
 import com.interrupt.server.member.entity.Member
 import com.interrupt.server.member.repository.MemberRepository
 import org.springframework.security.authentication.AuthenticationManager
